Update patch set 1
Patch Set 1:

You still need the AP and APN create/search/select options. As well the 2G only toggle I suppose. 

I updated our old overlay for Phone here to include the only mode we support that's missing (8) from the default settings.

http://review.cyanogenmod.org/#/c/44135/

The issue now is with signal strength. The fact we set the phone type as CDMA on LTE is the main cause. This results in the CDMA stack always passing !isGsm to the SignalStrength constructor. Which results in signal strength being read from CDMA vars in the parcel (which are obviously useless in GSM mode).

This patch is WIP and until we resolve the signal issues in the telephony stack, should NOT be merged. I does however allow the device to connect to GSM and WCDMA networks, utilizing voice, edge, and hspa+ techs. Just signal strength  is all derped up.

Patch-set: 1