Update patch set 2
Patch Set 2:

Ethan, before I push another patch that adds these missing stings, where would you prefer them?

We can't have another system.prop in the common folder so my best idea was to put it into a init file as 'setprop blah.cam 4mp'. It works and it would be in all m7's that way; I threw it in with init.qcom.rc, but I just don't want to waste anyone’s time pushing it in a patch that will have to be resubmitted.

Patch-set: 2