Update patch set 2
Patch Set 2: Code-Review+1

Markus, I removed PRODUCT_MODEL from the overrides because it is defined and correct (for CM purposes) in full_m7spr.mk. For whatever reason, I kept your re-arrangement of overrides below original definitions (but, do understand that it doesn't have any impact on the final build). If you do not believe this commit is representative of your work, let me know, and I will not submit it.

Patch-set: 2