Update patch set 3
Patch Set 3:

Upon adding this file, all the HTC apps in the Play Store become available, though many report that they are "incompatible with this device". I will see if this is rectifiable with a build.prop change, or something of the sort. If the apps can be made to work, I will upload a patchset with that info, otherwise, I'll abandon the change. There's no use in providing users apps they can't use.

Patch-set: 3