m7spr: Fix LTE and SIM
The m7spr reports "NO SIM CARD". This should fix this and is based on
changes made to our sister device, the t6spr. As well as fixing LTE.
This also puts our prop file a little more inline with what is done for
LTE and such on the stock Sense ROMs as well.

Change-Id: Ic899647ed96158f43dbeecb5c9510a8817859cd9