Update patch set 2
Patch Set 2:

I have now tested this change on my m7spr, and I can confirm that it resolves the issues with the SIM card being recognized. I, however, am not in an area with LTE coverage, so I am unable to 100% confirm that aspect of the change. However, I believe the two issues are interrelated.

Patch-set: 2