Update patch set 4
Patch Set 4: Code-Review-1

It's not that this _seems_ excessive as Ethan says, it _is_ excessive. Please take the time to narrow down the changes that have an effect on the issue(s) you're trying to resolve. Do not take t6spr as a template; I was not at all pleased when I saw the changes made to that device. Basically, when you're throwing darts, only keep the ones that actually hit the board.

Patch-set: 4