Update patch set 4
Patch Set 4:

Well, if there are still reports on other devices, either there is something CM-wide that we need to fix or we need to fix something in many CDMA device trees. I am not sure which that is, and I have no clue what to narrow it down to.

I'll compare similar build.prop files across different devices and try to discover which devices still suffer from APN/SIM/LTE issues, and see if there is a fix that works on them.

Patch-set: 4