Update patch set 5
Patch Set 5:

Its definitely better , but if you have a lock on your sim card sometimes it doesn't register that the pin when though , a simple back button gets the screen away . While others it works . It's a hit and miss . Data issues are not a problem unless you switch to anything other then LTE , such as CDMA Evo auto the APN's will disappear. So other then those little bugs this is a big improvement.

Patch-set: 5